requirements.txt
Flask
Flask-Script
Flask-Pymongo
